requirements.txt
python-telegram-bot==20.6
gspread==6.1.0
google-auth==2.23.0
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
pandas==2.2.3
python-dotenv==1.0.0
pillow==10.4.0
qrcode==7.4.2
Flask==2.3.3
gunicorn==20.1.0
setuptools==69.5.1
wheel==0.43.0